refactor v2 pipe tests
